Move zipper to snowpark instead of public api
