+++ src/snowcli/cli/snowpark/snowpark_shared.py
@@ -7,9 +7,9 @@ from typing import List
 import click
 import typer
 from requirements.requirement import Requirement
-from snowcli.api.utils.zipper import zip_dir
 from snowcli.cli.snowpark import package_utils
 from snowcli.cli.snowpark.models import PypiOption
+from snowcli.cli.snowpark.zipper import zip_dir
 
 PyPiDownloadOption: PypiOption = typer.Option(
     PypiOption.ASK.value, help="Whether to download non-Anaconda packages from PyPi."
