+++ src/snowcli/cli/snowpark/zipper.py
