+++ tests/test_zipper.py
@@ -1,6 +1,6 @@
 from zipfile import ZipFile
 
-from snowcli.api.utils.zipper import add_file_to_existing_zip, zip_dir
+from snowcli.cli.snowpark.zipper import add_file_to_existing_zip, zip_dir
 from snowcli.cli.snowpark.commands import _replace_handler_in_zip
 
 from tests.testing_utils.fixtures import *
